Move SmallSelectDialog helper to DeltaDialog
